Refactor frontend routing, secure API, and fix memory leaks.
- Renamed `src/ui/App.jsx` to `src/ui/Dashboard.jsx`.
- Created `src/ui/App.jsx` with `BrowserRouter` and `ProtectedRoute`.
- Updated API keys to use `REACT_APP_` prefix.
- Fixed memory leaks in `src/nlp/cd-detector.js` using `tf.tidy`, `dispose`, and `cosineProximity`.
- Added wildcard route in `server/index.js` for client-side routing.
- Enabled auth token interceptor in `src/services/api.js`.
- Implemented rate limiting in `server/index.js` to address security vulnerability.
- Fixed `webpack-dev-server` dependency conflict in `package.json`.
- Updated `SensitivitySelector.test.js` to use `fireEvent` for reliable MUI Select testing.
- Cleaned up package-lock.json to ensure consistent dependencies.
