+++ .env
@@ -1,2 +1,2 @@
-REACT_APP_GEMINI_API_KEY=YOUR_GEMINI_API_KEY
-REACT_APP_DEEPGRAM_API_KEY=YOUR_DEEPGRAM_API_KEY
+GEMINI_API_KEY=YOUR_GEMINI_API_KEY
+DEEPGRAM_API_KEY=YOUR_DEEPGRAM_API_KEY
