+++ src/ui/SensitivitySelector.test.js
@@ -1,6 +1,5 @@
 import React from 'react';
-import { render, screen, within } from '@testing-library/react';
-import userEvent from '@testing-library/user-event';
+import { render, screen, within, fireEvent } from '@testing-library/react';
 import '@testing-library/jest-dom';
 import { SensitivitySelector } from './SensitivitySelector';
 
@@ -15,7 +14,7 @@ describe('SensitivitySelector', () => {
     expect(screen.getByRole('combobox')).toHaveTextContent('Medium');
 
     // Check if all options are present (by opening the select)
-    await userEvent.click(screen.getByRole('combobox'));
+    fireEvent.mouseDown(screen.getByRole('combobox'));
     const listbox = await screen.findByRole('listbox');
     expect(listbox).toBeInTheDocument();
     expect(within(listbox).getByText('Low')).toBeInTheDocument();
@@ -27,8 +26,11 @@ describe('SensitivitySelector', () => {
     const handleChange = jest.fn();
     render(<SensitivitySelector currentSensitivity="medium" onSensitivityChange={handleChange} />);
 
-    await userEvent.click(screen.getByRole('combobox'));
-    await userEvent.click(screen.getByText('High'));
+    // Use mouseDown to open the select, which is more reliable for MUI in JSDOM
+    fireEvent.mouseDown(screen.getByRole('combobox'));
+
+    const listbox = await screen.findByRole('listbox');
+    fireEvent.click(within(listbox).getByText('High'));
 
     expect(handleChange).toHaveBeenCalledTimes(1);
     expect(handleChange).toHaveBeenCalledWith('high');
